requirements.txt
Flask_WTF==0.14.2
WTForms==2.2.1
Flask_SQLAlchemy==2.3.2
Flask==1.0.2
pytest==4.3.0
Pillow==5.4.1
secrets==1.0.2
